# doc/api.yaml
openapi: '3.0.2'
info:
  title: WASAPhoto API Specifications
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: "1.0"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for authentication in WASAPhoto.
      bearerFormat: '^[a-z0-9]{64}$'
  schemas:
    ID:
      title: ID
      description: |-
        Autoincrement integer field used for identify a particular resource (a comment, a post, ecc.)
      type: integer
      example: 795739
    Username:
      title: Username
      description: Username of a user
      type: string
      example: "chri_genna02"
      minLength: 8
      maxLength: 16
      pattern: '^[a-zA-Z0-9_-]{8,16}$'
    User:
      title: User
      description: |-
        A user in WASAPhoto is represented by a unique ID, its unique username, name and birthdate.
      properties:
        uuid:
          $ref: '#/components/schemas/ID'
        username:
          $ref: '#/components/schemas/Username'
        profilePic:
          $ref: "#/components/schemas/Photo"
        name:
          description: Real-life name/surname of the user
          type: string
          pattern: ''
          minLength: 2
          maxLength: 32
          example: Christian Gennarelli 
        birthdate:
          $ref: '#/components/schemas/Date'
    UserList:
      title: UserList
      description: |-
        Collection of users. 
        A maximum number of 16 users is returned, according to some backend principle (example: datetime)
      type: array
      items:
        $ref: '#/components/schemas/User'  
      minItems: 0
      maxItems: 16 
    Photo:
      title: Photo
      format: binary
      type: string
      description: |-
        Binary representation of a photo, wrapped in a base64-encoded string. 
        Only multiple profile pics are sent via base64 for efficiency reasons.
        URL pointing to post images will be first sent, and then the client will have the ability
        to obtain the photo via a separate endpoint, i.e. GetPhotoFromURL.
      minLength: 0
      maxLength: 2000000 # 2MB photo
      example: iVBORw0KGgbW8gY2hyaXN0aWFuCghyaXN0aWFuCgw0KGgbW8gYyaXN0aWFuCgw0KGgbW8gW8gYyaXN0aW8gYyaXN0aWF
...
    Date:
      title: date
      description: Date.
      example: "2023-11-29"
      type: string
      format: date
      pattern: '^([0-9]{4})-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$'
      minLength: 10
      maxLength: 10  
...
    Error:
      title: Error
      description: |-
        Schema defining what is an error when returned in response of a HTTP requests.
      properties:
        code: # Code returned by the server
          type: string
          description: Status code returned directly by the server(s).
          example: "404"
          minLength: 3
          maxLength: 3
        description:  
          type: string
          description: |- 
            Description of the error in human-readable language.
          example: 'Username not found. Please check that the username is correct, and try again.'
          minLength: 1
          maxLength: 128
          pattern: ''
paths:
...
  /users/:
    parameters:
      - in: query
        name: searched_username
        description: Username that the authenticated user is trying to find.
        schema:
          $ref: '#/components/schemas/Username'
    get:
      operationId: searchUser
      tags: ['USER']
      summary: Search for a user
      description: |-
        A particular user - here, "username" - can research another user - here, "searched-user" - 
        or a set of users by their usernames matching the substring provided.
      responses:
        '200': # OK
          description: List of returned usernames matching the provided substring.
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/UserList'
        '204': # Content not found
          description: No user has been found with such a substring in their usernames.
        '400': # Bad request provided
          description: The request cannot be processed.
          content:  
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Unauthenticated
          description: The client has not been authenticated (yet).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404': # Content not found
          description: Provided username not found. 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error.
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...